Completed Value Added Features and Refactors

+++ HotelListing.API/Controllers/CountriesV2Controller.cs
@@ -1,26 +1,28 @@
 ﻿using AutoMapper;
 using HotelListing.API.Core.Contracts;
-using HotelListing.API.Core.Models;
+using HotelListing.API.Core.Exceptions;
 using HotelListing.API.Core.Models.Country;
+using HotelListing.API.Data;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OData.Query;
 using Microsoft.EntityFrameworkCore;
 
 namespace HotelListing.API.Controllers;
 
 [Route("api/v{version:apiVersion}/countries")]
 [ApiController]
-[ApiVersion("1.0", Deprecated = true)]
-public class CountriesController : ControllerBase
+[ApiVersion("2.0")]
+public class CountriesV2Controller : ControllerBase
 {
   private readonly ICountriesRepository _countriesRepository;
-  private readonly ILogger<CountriesController> _logger;
+  private readonly ILogger<CountriesV2Controller> _logger;
   private readonly IMapper _mapper;
 
-  public CountriesController(
+  public CountriesV2Controller(
     ICountriesRepository countriesRepository,
     IMapper mapper,
-    ILogger<CountriesController> logger
+    ILogger<CountriesV2Controller> logger
   )
   {
     _countriesRepository = countriesRepository;
@@ -28,43 +30,33 @@ public class CountriesController : ControllerBase
     _logger = logger;
   }
 
-  // GET: api/Countries/GetAll
-  [HttpGet("GetAll")]
-  [Authorize]
+  // GET: api/Countries
+  [HttpGet]
+  [EnableQuery]
   public async Task<ActionResult<IEnumerable<GetCountryDto>>> GetCountries()
   {
     _logger.LogInformation($"Action: {nameof(GetCountries)}");
 
-    var countries = await _countriesRepository.GetAllAsync<GetCountryDto>();
+    var countries = await _countriesRepository.GetAllAsync();
 
-    return Ok(countries);
-  }
+    var getCountries = _mapper.Map<List<GetCountryDto>>(countries);
 
-  // GET: api/Countries/?StartIndex=0&PageSize=25&PageNumber=1
-  [HttpGet]
-  [Authorize]
-  public async Task<ActionResult<PagedResult<GetCountryDto>>> GetPagedCountries(
-    [FromQuery] QueryParameters queryParameters
-  )
-  {
-    _logger.LogInformation($"Action: {nameof(GetPagedCountries)}");
-
-    var pagedCountriesResult =
-      await _countriesRepository.GetAllAsync<GetCountryDto>(queryParameters);
-
-    return Ok(pagedCountriesResult);
+    return Ok(getCountries);
   }
 
   // GET: api/Countries/5
   [HttpGet("{id}")]
-  [Authorize]
   public async Task<ActionResult<CountryDto>> GetCountry(int id)
   {
     _logger.LogInformation($"Action: {nameof(GetCountry)} by id: {id}");
 
     var country = await _countriesRepository.GetDetails(id);
 
-    return Ok(country);
+    if (country == null) throw new NotFoundException(nameof(GetCountry), id);
+
+    var countryDto = _mapper.Map<CountryDto>(country);
+
+    return Ok(countryDto);
   }
 
   // PUT: api/Countries/5
@@ -79,9 +71,16 @@ public class CountriesController : ControllerBase
 
     if (id != updateCountryDto.Id) return BadRequest("Invalid Record Id");
 
+    var country = await _countriesRepository.GetAsync(id); // tracked
+
+    if (country == null)
+      throw new NotFoundException(nameof(PutCountry), id);
+
+    _mapper.Map(updateCountryDto, country);
+
     try
     {
-      await _countriesRepository.UpdateAsync(id, updateCountryDto);
+      await _countriesRepository.UpdateAsync(country); // change entity
     }
     catch (DbUpdateConcurrencyException)
     {
@@ -103,16 +102,11 @@ public class CountriesController : ControllerBase
       $"Action: {nameof(PostCountry)} with name: {createCountry.Name}"
     );
 
-    var country =
-      await _countriesRepository.AddAsync<CreateCountryDto, GetCountryDto>(
-        createCountry
-      );
+    var country = _mapper.Map<Country>(createCountry);
 
-    return CreatedAtAction(
-      nameof(GetCountry),
-      new { id = country.Id },
-      country
-    );
+    await _countriesRepository.AddAsync(country);
+
+    return CreatedAtAction("GetCountry", new { id = country.Id }, country);
   }
 
   // DELETE: api/Countries/5
@@ -121,6 +115,10 @@ public class CountriesController : ControllerBase
   public async Task<IActionResult> DeleteCountry(int id)
   {
     _logger.LogInformation($"Action: {nameof(DeleteCountry)} with id: {id}");
+
+    var country = await _countriesRepository.GetAsync(id);
+    if (country == null)
+      throw new NotFoundException(nameof(DeleteCountry), id);
 
     await _countriesRepository.DeleteAsync(id);
